requirements.txt
aiosqlite==0.17.0
asgiref==3.4.1
click==8.0.1
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.68.1
h11==0.12.0
iso8601==0.1.16
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
